修复 JavaScriptProcessor Replacer 可能返回空字符串的 bug，修复失效测试

+++ src/JavaScriptProcessor/Replacer/FileReplacer.ts
@@ -18,10 +18,13 @@ export class FileReplacer
 
     public async doProcess(): Promise<void>
     {
-        const {filePath} = this.scriptFile;
-        const fileContent = await fse.promises.readFile(filePath, this.encoding);
-        const stringReplacer = new StringReplacer(fileContent, this.replaceInfos);
-        const processedFileContent = stringReplacer.getReplacedString();
-        await fse.outputFile(filePath, processedFileContent, this.encoding);
+        if(this.replaceInfos.length > 0)
+        {
+            const {filePath} = this.scriptFile;
+            const fileContent = await fse.promises.readFile(filePath, this.encoding);
+            const stringReplacer = new StringReplacer(fileContent, this.replaceInfos);
+            const processedFileContent = stringReplacer.getReplacedString();
+            await fse.outputFile(filePath, processedFileContent, this.encoding);
+        }
     }
 }
